solicitud/requirements.txt
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
cryptography==41.0.4
cssselect2==0.7.0
Flask==2.2.2
fonttools==4.44.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==10.1.0
protobuf==3.20.1
pycparser==2.21
pydyf==0.8.0
PyMySQL==1.1.0
pyphen==0.14.0
python-dotenv==0.21.0
reportlab==4.0.6
requests==2.31.0
six==1.16.0
soupsieve==2.5
tinycss2==1.2.1
urllib3==2.0.7
weasyprint==60.1
webencodings==0.5.1
Werkzeug==2.2.2
zopfli==0.2.3
